centralize peer discovery a bit more
