+++ mesh_client.go
@@ -8,6 +8,8 @@ import (
 	"net/http"
 	"strconv"
 	"time"
+
+	"tailscale.com/tsnet"
 )
 
 // MeshClient handles authenticated HTTP communication over the Tailscale mesh.
@@ -37,6 +39,41 @@ func (m *MeshClient) signRequest(req *http.Request) {
 	req.Header.Set(HeaderNaraName, m.name)
 	req.Header.Set(HeaderNaraTimestamp, strconv.FormatInt(timestamp, 10))
 	req.Header.Set(HeaderNaraSignature, m.keypair.SignBase64([]byte(message)))
+}
+
+// BuildMeshURL builds a mesh URL for a given IP and path
+func BuildMeshURL(ip string, path string) string {
+	return fmt.Sprintf("http://%s:%d%s", ip, DefaultMeshPort, path)
+}
+
+// DiscoverMeshPeers discovers peers using tsnet Status API
+func DiscoverMeshPeers(ctx context.Context, server *tsnet.Server) ([]TsnetPeer, error) {
+	lc, err := server.LocalClient()
+	if err != nil {
+		return nil, fmt.Errorf("failed to get local client: %w", err)
+	}
+
+	status, err := lc.Status(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get status: %w", err)
+	}
+
+	var peers []TsnetPeer
+	for _, peer := range status.Peer {
+		if len(peer.TailscaleIPs) == 0 || peer.HostName == "" {
+			continue
+		}
+		// Skip backup tool peers
+		if len(peer.HostName) >= 12 && peer.HostName[:12] == "nara-backup-" {
+			continue
+		}
+		peers = append(peers, TsnetPeer{
+			Name: peer.HostName,
+			IP:   peer.TailscaleIPs[0].String(),
+		})
+	}
+
+	return peers, nil
 }
 
 // FetchSyncEvents fetches events from a peer via POST /events/sync
