+++ transport_mesh_helpers.go
@@ -43,110 +43,3 @@ func ConnectToMesh(ctx context.Context, config MeshConnectionConfig) (*tsnet.Ser
 
 	return server, server.HTTPClient(), nil
 }
-
-// SignMeshRequest adds mesh authentication headers to an HTTP request
-// Uses the provided name and keypair to sign
-func SignMeshRequest(req *http.Request, name string, keypair NaraKeypair) {
-	timestamp := time.Now().UnixMilli()
-	message := fmt.Sprintf("%s%d%s%s", name, timestamp, req.Method, req.URL.Path)
-
-	req.Header.Set(HeaderNaraName, name)
-	req.Header.Set(HeaderNaraTimestamp, strconv.FormatInt(timestamp, 10))
-	req.Header.Set(HeaderNaraSignature, keypair.SignBase64([]byte(message)))
-}
-
-// DiscoverMeshPeers discovers peers using tsnet Status API
-func DiscoverMeshPeers(ctx context.Context, server *tsnet.Server) ([]TsnetPeer, error) {
-	lc, err := server.LocalClient()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get local client: %w", err)
-	}
-
-	status, err := lc.Status(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get status: %w", err)
-	}
-
-	var peers []TsnetPeer
-	for _, peer := range status.Peer {
-		if len(peer.TailscaleIPs) == 0 || peer.HostName == "" {
-			continue
-		}
-		// Skip backup tool peers
-		if len(peer.HostName) >= 12 && peer.HostName[:12] == "nara-backup-" {
-			continue
-		}
-		peers = append(peers, TsnetPeer{
-			Name: peer.HostName,
-			IP:   peer.TailscaleIPs[0].String(),
-		})
-	}
-
-	return peers, nil
-}
-
-// BuildMeshURL builds a mesh URL for a given IP and path
-func BuildMeshURL(ip string, path string) string {
-	return fmt.Sprintf("http://%s:%d%s", ip, DefaultMeshPort, path)
-}
-
-// FetchSyncEventsFromMesh makes a mesh-authenticated request to /events/sync
-// Returns the events and whether the response was signature-verified
-// This is used by both boot recovery and the backup tool
-func FetchSyncEventsFromMesh(
-	client *http.Client,
-	fromName string,
-	keypair NaraKeypair,
-	peerIP string,
-	peerName string,
-	subjects []string,
-	sliceIndex int,
-	sliceTotal int,
-	maxEvents int,
-) ([]SyncEvent, bool) {
-	// Build request
-	reqBody := SyncRequest{
-		From:       fromName,
-		Subjects:   subjects,
-		SinceTime:  0, // get all events
-		SliceIndex: sliceIndex,
-		SliceTotal: sliceTotal,
-		MaxEvents:  maxEvents,
-	}
-
-	jsonBody, err := json.Marshal(reqBody)
-	if err != nil {
-		return nil, false
-	}
-
-	// Make HTTP request to neighbor's mesh endpoint
-	url := BuildMeshURL(peerIP, "/events/sync")
-	req, err := http.NewRequest("POST", url, bytes.NewReader(jsonBody))
-	if err != nil {
-		return nil, false
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// Add mesh authentication headers (Ed25519 signature)
-	SignMeshRequest(req, fromName, keypair)
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, false
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, false
-	}
-
-	// Parse response
-	var response SyncResponse
-	if err := json.NewDecoder(resp.Body).Decode(&response); err != nil {
-		return nil, false
-	}
-
-	// TODO: Add response signature verification (requires access to neighbourhood public keys)
-	// For now, return false for verified (caller can decide whether to trust)
-	return response.Events, false
-}
